Removed the dependency of the microservice communication class with the requests microservice.
Added a new GET mapping to getCredits/ that will return user's credits instead of parsing a whole user object in the transactions microservice.

There was a problem with gradle bootRun only if I referenced the requests microservice as a project(). If I used the compiled .jar (but then gitlab CI failed, as it could not find the file), gradle bootRun worked as expected. Also, gradle transactions:bootRun always worked, which leads me to believe that there is an error with gradle.