Extracted some of the JwtFilter logic to three new authentication helper classes.
This reduced the coupling of JwtFilter from 'low-medium' to 'low'. It also reduced the attemtAuthentication() method's coupling from 'low-medium' to 'low'.

This refactoring made the coupling of the helpers package too high though, so I refactored the package into two smaller packages - authentication and communication.

The Authentication helpers are responsible for registering and authenticating a user, while the Communication helpers are responsible for the microservice communication between the authentication and requests microservices.